feat: add aliases from the old CLI and splitted get commands in separate files
